Move around some tests
