+++ py-polars/tests/unit/streaming/test_streaming_unique.py
@@ -4,6 +4,7 @@ from typing import Any
 import pytest
 
 import polars as pl
+from polars.testing import assert_frame_equal
 
 pytestmark = pytest.mark.xdist_group("streaming")
 
@@ -30,3 +31,18 @@ def test_streaming_out_of_core_unique(
     # TODO: Re-enable this check when this issue is fixed: https://github.com/pola-rs/polars/issues/10466
     _ = capfd.readouterr().err
     # assert "OOC groupby started" in err
+
+
+def test_streaming_unique(monkeypatch: Any, capfd: Any) -> None:
+    monkeypatch.setenv("POLARS_VERBOSE", "1")
+    df = pl.DataFrame({"a": [1, 2, 2, 2], "b": [3, 4, 4, 4], "c": [5, 6, 7, 7]})
+    q = df.lazy().unique(subset=["a", "c"], maintain_order=False).sort(["a", "b", "c"])
+    assert_frame_equal(q.collect(streaming=True), q.collect(streaming=False))
+
+    q = df.lazy().unique(subset=["b", "c"], maintain_order=False).sort(["a", "b", "c"])
+    assert_frame_equal(q.collect(streaming=True), q.collect(streaming=False))
+
+    q = df.lazy().unique(subset=None, maintain_order=False).sort(["a", "b", "c"])
+    assert_frame_equal(q.collect(streaming=True), q.collect(streaming=False))
+    (_, err) = capfd.readouterr()
+    assert "df -> re-project-sink -> sort_multiple" in err
